K 23 apr   2025 12:15:29
Add DOTOOL_PIPE to session environment

The dotool client uses the DOTOOL_PIPE environment variable to locate the daemon's communication pipe.

This commit adds DOTOOL_PIPE to `environment.sessionVariables` so it is automatically set for users logging in, simplifying client setup.

It also introduces a variable for the pipe path definition to ensure consistency between the daemon's configuration and the value set in the user's environment.