P 04 mai 2025 16:57:24
Disable the sherlock systemd user service and the `daemonize` option in its configuration file. This change is made because daemonizing currently prevents sherlock from receiving input via stdin, as noted in the configuration.

Remove the keybindings for the `albert` and `fuzzel` launchers in the niri configuration. Sherlock remains bound to Mod+S as the primary launcher.

Additionally, disable naggin CSS file warning for sherlock by setting the `css` path to `/dev/null`, and add more applications to the `exclude` list in the fallback configuration.